Windows DNS Server Remote Code Execution Vulnerability (CVE-2021-26897)
A remote code execution vulnerability, identified as CVE-2021-26897, exists in Windows Domain Name System (DNS) servers. An attacker who successfully exploited the vulnerability could run arbitrary code in the context of the Local System Account over the network. To be vulnerable, the DNS server would need to have dynamic updates enabled. The vulnerability is rated with a CVSv3 score of 9.8/8.5.